refactor: moving logic out of attestation env tests to celotool. Did not realise it was duplicated
