+++ packages/env-tests/src/tests/attestation.ts
@@ -1,3 +1,10 @@
+import {
+  getIdentifierAndPepper,
+  pollForMessagesAndCompleteAttestations,
+  reportErrors,
+  requestAttestationsFromIssuers,
+  requestMoreAttestations,
+} from '@celo/celotool/src/lib/attestation'
 import { AttestationsWrapper } from '@celo/contractkit/lib/wrappers/Attestations'
 import { StableTokenWrapper } from '@celo/contractkit/lib/wrappers/StableTokenWrapper'
 import { generateKeys, generateMnemonic } from '@celo/utils/lib/account'
@@ -12,14 +19,7 @@ import { IncomingPhoneNumberInstance } from 'twilio/lib/rest/api/v2010/account/i
 import { EnvTestContext } from '../context'
 import { envVar, fetchEnv } from '../env'
 import { getKey, ONE, TestAccounts } from '../scaffold'
-import {
-  getIdentifierAndPepper,
-  getPhoneNumber,
-  pollForMessagesAndCompleteAttestations,
-  reportErrors,
-  requestAttestationsFromIssuers,
-  requestMoreAttestations,
-} from '../shared/attestation'
+import { getPhoneNumber } from '../shared/attestation'
 
 async function fundAttestationAccount(
   context: EnvTestContext,
@@ -123,6 +123,7 @@ export function runAttestationTest(
 
       await pollForMessagesAndCompleteAttestations(
         attestations,
+        // @ts-ignore
         twilioClient,
         phoneNumber.phoneNumber,
         identifier,
@@ -176,6 +177,7 @@ export function runAttestationTest(
 
       await pollForMessagesAndCompleteAttestations(
         attestations,
+        // @ts-ignore
         twilioClient,
         phoneNumber.phoneNumber,
         identifier,
